# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /crm/workspaces:
    get:
      operationId: listCrmWorkspaces
      tags: [crm]
      summary: List workspaces available to the current user
      responses:
        "200":
          description: User workspaces and current selection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CrmWorkspaceList" }
    post:
      operationId: createCrmWorkspace
      tags: [crm]
      summary: Create an agency workspace
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CrmWorkspaceInput" }
      responses:
        "201":
          description: Created agency workspace
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CrmWorkspace" }
  /crm/workspaces/{workspaceId}/members:
    get:
      operationId: listCrmWorkspaceMembers
      tags: [crm]
      summary: List members and roles in a workspace
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Workspace members
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CrmWorkspaceMember" }
    post:
      operationId: inviteCrmWorkspaceMember
      tags: [crm]
      summary: Send a verified-email-bound workspace invitation
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CrmWorkspaceInvitationInput" }
      responses:
        "201":
          description: Invitation email sent with the one-time acceptance code
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CrmWorkspaceInvitation" }
  /crm/workspaces/{workspaceId}/members/{membershipId}:
    patch:
      operationId: updateCrmWorkspaceMember
      tags: [crm]
      summary: Change a workspace member's role
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema: { type: integer }
        - in: path
          name: membershipId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CrmWorkspaceMemberUpdate" }
      responses:
        "200":
          description: Updated workspace member
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CrmWorkspaceMember" }
    delete:
      operationId: removeCrmWorkspaceMember
      tags: [crm]
      summary: Remove a member from a workspace
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema: { type: integer }
        - in: path
          name: membershipId
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Member removed }
  /crm/workspace-invitations/accept:
    post:
      operationId: acceptCrmWorkspaceInvitation
      tags: [crm]
      summary: Accept an invitation for the signed-in verified email
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CrmWorkspaceInvitationAcceptInput" }
      responses:
        "200":
          description: Joined workspace
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CrmWorkspace" }
components:
  schemas:
...
    CrmWorkspace:
      type: object
      required: [id, name, kind, role, memberCount, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        kind: { type: string, enum: [personal, agency] }
        role: { type: string, enum: [owner, admin, agent, viewer] }
        memberCount: { type: integer }
        createdAt: { type: string, format: date-time }
    CrmWorkspaceList:
      type: object
      required: [currentWorkspaceId, workspaces]
      properties:
        currentWorkspaceId: { type: integer }
        workspaces:
          type: array
          items: { $ref: "#/components/schemas/CrmWorkspace" }
    CrmWorkspaceInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 2, maxLength: 80 }
    CrmWorkspaceMember:
      type: object
      required: [id, userId, email, role, joinedAt, isCurrentUser]
      properties:
        id: { type: integer }
        userId: { type: string }
        email: { type: ["string", "null"] }
        role: { type: string, enum: [owner, admin, agent, viewer] }
        joinedAt: { type: string, format: date-time }
        isCurrentUser: { type: boolean }
    CrmWorkspaceInvitationInput:
      type: object
      required: [email, role]
      properties:
        email: { type: string, format: email }
        role: { type: string, enum: [admin, agent, viewer] }
    CrmWorkspaceInvitation:
      type: object
      required: [email, role, code, expiresAt]
      properties:
        email: { type: string, format: email }
        role: { type: string, enum: [admin, agent, viewer] }
        code: { type: string }
        expiresAt: { type: string, format: date-time }
    CrmWorkspaceMemberUpdate:
      type: object
      required: [role]
      properties:
        role: { type: string, enum: [admin, agent, viewer] }
    CrmWorkspaceInvitationAcceptInput:
      type: object
      required: [code]
      properties:
        code: { type: string, minLength: 20 }