# lib/api-zod/src/generated/types/crmWorkspaceInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CrmWorkspaceInput {
  /**
     * @minLength 2
     * @maxLength 80
     */
  name: string;
}