# lib/api-zod/src/generated/types/crmWorkspaceInvitationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CrmWorkspaceInvitationInputRole } from './crmWorkspaceInvitationInputRole';

export interface CrmWorkspaceInvitationInput {
  email: string;
  role: CrmWorkspaceInvitationInputRole;
}