# lib/api-zod/src/generated/types/crmWorkspaceMember.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CrmWorkspaceMemberRole } from './crmWorkspaceMemberRole';

export interface CrmWorkspaceMember {
  id: number;
  userId: string;
  /** @nullable */
  email: string | null;
  role: CrmWorkspaceMemberRole;
  joinedAt: Date;
  isCurrentUser: boolean;
}