# lib/api-zod/src/generated/types/crmWorkspaceMemberUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CrmWorkspaceMemberUpdateRole } from './crmWorkspaceMemberUpdateRole';

export interface CrmWorkspaceMemberUpdate {
  role: CrmWorkspaceMemberUpdateRole;
}